Update (some) callers to follow more strict span rules
- allows us to migrate to std::span, keeping SkSpan (for now)
  as an alias.

- idea is to land call-site-cleanups separately, so that when we
  do try to actually throw the switch, that CL only contains the
  switch (build-flag) itself.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1097896
Commit-Queue: Mike Reed <mike@reedtribe.org>

+++ src/base/SkFixedArray.h
@@ -52,12 +52,12 @@ public:
     }
 
     FixedArray(const FixedArray<N, T>& that) {
-        this->reset({that.data(), that.size()});
+        this->reset({that.data(), (size_t)that.size()});
     }
 
     FixedArray<N, T>& operator=(const FixedArray<N, T>& that) {
         if (this != &that) {
-            this->reset({that.data(), that.size()});
+            this->reset({that.data(), (size_t)that.size()});
         }
         return *this;
     }
